initial commit: Hw7

+++ src/main/java/edu/hw7/Task2.java
@@ -5,9 +5,9 @@ import java.util.List;
 
 public class Task2 {
 
-    public static long parallelFactorial(int value) {
+    public static int parallelFactorial(int value) {
         List<Integer> factorialSequence = new ArrayList<>();
-        for(int i = 1; i - 1< value; i++){
+        for (int i = 1; i - 1 < value; i++) {
             factorialSequence.add(i);
         }
         return factorialSequence.parallelStream().reduce(1, (a, b) -> a * b);
